+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - **BREAKING**: Don't merge framework configuration.
 
+### Removed
+- **BREAKING**: Middleware ValidateCsrfToken.
+
 ## [0.27.0] - 2024-05-31
 
 ### Added